Download all gsets from msigdb for a species

Download all gsets from msigdb for a species

msigdb_download_all(species = "Homo sapiens", output_dir = NULL)

Arguments

species
A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)

output_dir
A directory path where gene sets are downloaded instead of to the package location

Value

A list containing the output directory and version number of gene sets

Examples

msigdb_path <- msigdb_download_all("Homo sapiens")

#> v6.2.1
#> Downloading genesets to...
#> /var/folders/7j/bg701dss38lc2xq26vt_by240000gn/T//RtmpHOO3hV
#> - C1 -> 326 Gene Sets 
#> - C2.CGP -> 3433 Gene Sets 
#> - C2.CP -> 252 Gene Sets 
#> - C2.CP.BIOCARTA -> 217 Gene Sets 
#> - C2.CP.KEGG -> 186 Gene Sets 
#> - C2.CP.REACTOME -> 674 Gene Sets 
#> - C3.MIR -> 221 Gene Sets 
#> - C3.TFT -> 615 Gene Sets 
#> - C4.CGN -> 427 Gene Sets 
#> - C4.CM -> 431 Gene Sets 
#> - C5.BP -> 4436 Gene Sets 
#> - C5.CC -> 580 Gene Sets 
#> - C5.MF -> 901 Gene Sets 
#> - C6 -> 189 Gene Sets 
#> - C7 -> 4872 Gene Sets 
#> - H -> 50 Gene Sets