A relational genesets object

A relational genesets object

A relational genesets object

See Also

gsets

Public fields

genesets

A list of genesets where list names refers to geneset labels and values are geneset members represented as a vector

nodes

A data frame of labeled nodes

edges

A data frame of directed edges

name

A character vector describing source of genesets

version

A character vector describing versioning

Methods

Public methods

rgsets$new()

rgsets$print()

rgsets$info()

rgsets$reduce()

rgsets$subset()

rgsets$clone()

Method new()

Create a rgsets object

Usage

rgsets$new(
  genesets,
  nodes,
  edges,
  name = "Custom",
  version = "",
  quiet = FALSE
)

Arguments

genesets

A list of genesets where list names refers to geneset labels and values are geneset members represented as a vector

nodes

A data frame of labeled nodes

edges

A data frame of directed edges

name

A character vector describing source of genesets

version

A character vector describing versioning

quiet

Use true to silence warnings

Returns

A new rgsets object

Method print()

Print relational genesets information

Usage

rgsets$print()

Returns

NULL

Method info()

Returns versioning information

Usage

rgsets$info()

Returns

A character vector with name and version

Method reduce()

Reduces genesets to a background distribution of symbols

Usage

rgsets$reduce(background)

Arguments

background

A character vector of symbols

Returns

A rgsets object

Method subset()

Subsets genesets on a character vector of labels

Usage

rgsets$subset(labels)

Arguments

labels

A character vector of genesets

Returns

A rgsets object

Method clone()

The objects of this class are cloneable with this method.

Usage

rgsets$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

testdat <- readRDS(file.path(system.file("extdata", package="hypeR"), "testdat.rds"))
rgsets <- rgsets$new(genesets=testdat$genesets, nodes=testdat$nodes, edges=testdat$edges, 
                     name="Example", version="v1.0")